Add support for Wix 6 generated DVD image in version 270 and later. (#579)
The DVD image for BC version 2.7 is created with WIX 6, and uses
tokenized names for standard folders. All relevant scripts has been
modified to use functions for path names related to the DVD image. The
functions will evaluate the current layout and return the correct root
path for the uncompressed files.
The 270 subfolder is based on the existing 260 folder with the addition
of path functions.

Co-authored-by: Niels-Henrik Sejthen <Niels-Henrik.Sejthen@microsoft.com>